Separate ScheduledTaskListCommandTest
